Silence warning for uint32_t to pointer cast
Summary: Silence warning for uint32_t to pointer cast.

Reviewed By: lavenzg

Differential Revision: D89912023

fbshipit-source-id: dc1b181fdf4016bcf2976f2b00fda9548b536000

+++ include/hermes/VM/JSWeakMapImpl.h
@@ -25,8 +25,8 @@ namespace detail {
 /// Uses invalid locations as slots for empty and tombstone keys.
 class WeakRefKey {
   /// Used to construct Empty/Tombstone WeakRefKey.
-  static constexpr uint32_t kEmptyKey = 0;
-  static constexpr uint32_t kTombstoneKey = 1;
+  static constexpr uintptr_t kEmptyKey = 0;
+  static constexpr uintptr_t kTombstoneKey = 1;
 
   /// Slot to the key object (packaged with its mapped value and owning map).
   WeakMapEntrySlot *slot_;
